+++ NEWS.md
@@ -1,10 +1,12 @@
 # dbscan 1.1-10-1 (2022-xx-xx)
 
 ## New Features
+* dbscan now retains information on distance method and border points.
 * HDBSCAN now supports long vectors to work with larger distance matrices. 
 * conversion from dist to kNN and frNN is now more memory efficient. It does no longer 
   coerce the dist object into a matrix of double the size, but extract the distances directly
   from the dist object.
+* Better description of how predict uses only Euclidean distances and more error checking.
 
 # dbscan 1.1-10 (2022-01-14)